--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -81,6 +81,7 @@
 - PR #5491 allow build libcudf in custom dir
 - PR #5503 Change `unique_count` to `distinct_count`
 - PR #5496 Rename .cu tests (zero cuda kernels) to .cpp files
+- PR #5518 split iterator and gather tests to speedup build tests
 
 ## Bug Fixes