Fixes #5464 
gather_test.cu (88s)
iterator_test.cu compilation time is 339s (5.65 minutes) ! This will be higher after Duration Types are added to FixedWidthTypes. 

- [x] splits iterator tests (339s to 83s+136s+142s)
- [x] splits gather_tests  (88s to 17s + 26s + 30s)