@harrism 2020-06-24
Have you done a ninja profile to show the benefit of this? (Total time is not reduced, but parallelism is increased, right?)