@karthikeyann 2020-06-24
> Have you done a ninja profile to show the benefit of this? (Total time is not reduced, but parallelism is increased, right?)

I did. I have attached the tracing json files of old and new builds [fresh_build.ninja_trace_old_new.zip](https://github.com/rapidsai/cudf/files/4824631/fresh_build.ninja_trace_old_new.zip) (view json using [chrome://tracing/](chrome://tracing/) )

Yes. But parallel builds gives benefit.

![image](https://user-images.githubusercontent.com/6488848/85537239-6ad1ab80-b631-11ea-89ee-2bf64562574e.png)