@harrism 2020-06-24
Nice job. I actually started splitting up the iterator tests for the same reason a while back but got distracted.   I think reductions will soon be a problem. :)