@karthikeyann 2020-06-24
After adding duration types, the compile time is too high. So I looked into it.
For reductions, I will try device side dispatch.

--
Karthikeyan Natarajan

________________________________
From: Mark Harris <notifications@github.com>
Sent: Wednesday, June 24, 2020 4:04:27 PM
To: rapidsai/cudf <cudf@noreply.github.com>
Cc: Karthikeyan Natarajan <knataraj@nvidia.com>; State change <state_change@noreply.github.com>
Subject: Re: [rapidsai/cudf] [REVIEW] split iterator and gather tests to speedup build tests (#5518)


Nice job. I actually started splitting up the iterator tests for the same reason a while back but got distracted. I think reductions will soon be a problem. :)

—
You are receiving this because you modified the open/close state.
Reply to this email directly, view it on GitHub<https://github.com/rapidsai/cudf/pull/5518#issuecomment-648740165>, or unsubscribe<https://github.com/notifications/unsubscribe-auth/ABRQGEBWHIZQ6OVJOQ7M66LRYHJDHANCNFSM4OC2XHNQ>.